+++ services/director-v2/src/simcore_service_director_v2/utils/computations.py
@@ -12,7 +12,7 @@ from ..models.comp_tasks import CompTaskAtDB
 from ..modules.catalog import CatalogClient
 from ..modules.db.tables import NodeClass
 
-log = logging.getLogger(__name__)
+_logger = logging.getLogger(__name__)
 
 _COMPLETED_STATES = (RunningState.ABORTED, RunningState.FAILED, RunningState.SUCCESS)
 _RUNNING_STATES = (RunningState.STARTED,)
